extract middleware to a separate package
